hasBatchim
