+++ docs/src/pages/docs/api/hasBatchim.en.mdx
@@ -1,0 +1,18 @@
+# hasBatchim
+Checks if the last character of a Korean string has a batchim (jongseong).
+
+```typescript
+hasBatchim(
+  str: string, 
+  // checking for single batchim 
+  options?: { single?: boolean }
+): boolean
+
+ 
+ ```typescript
+ hasBatchim('값') // true
+ hasBatchim('토') // false
+ hasBatchim('갑', { single: true }) // true
+ hasBatchim('값', { single: true }) // false
+ hasBatchim('토', { single: true }) // false
+ ````
